# specs/002-todo-backend-crud/contracts/todo-api-openapi.yaml
openapi: 3.0.3
info:
  title: Todo Application API
  description: API for managing todo tasks
  version: 1.0.0
  contact:
    name: API Support
    url: http://www.example.com/support
    email: support@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.todoapp.com/v1
    description: Production server
paths:
  /tasks:
    get:
      summary: Get all tasks
      description: Retrieve a list of all tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new task
      description: Create a new task with the provided details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to retrieve
      responses:
        '200':
          description: The requested task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a task
      description: Update an existing task with the provided details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a task
      description: Delete a specific task by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task to delete
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: Brief title or summary of the task
          minLength: 1
          maxLength: 255
          example: "Complete project proposal"
        description:
          type: string
          description: Detailed description of the task
          maxLength: 1000
          example: "Finish writing the project proposal document and send it to the manager by EOD."
        completed:
          type: boolean
          description: Flag indicating if the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2026-01-11T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: "2026-01-11T10:00:00Z"
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Brief title or summary of the task
          minLength: 1
          maxLength: 255
          example: "Complete project proposal"
        description:
          type: string
          description: Detailed description of the task
          maxLength: 1000
          example: "Finish writing the project proposal document and send it to the manager by EOD."
        completed:
          type: boolean
          description: Flag indicating if the task is completed
          default: false
          example: false
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: Brief title or summary of the task
          minLength: 1
          maxLength: 255
          example: "Complete project proposal"
        description:
          type: string
          description: Detailed description of the task
          maxLength: 1000
          example: "Finish writing the project proposal document and send it to the manager by EOD."
        completed:
          type: boolean
          description: Flag indicating if the task is completed
          example: true
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: A human-readable description of the error
          example: "Task not found"